pyproject.toml
[tool.poetry]
name = "mydatasales-back-end"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.taskipy.tasks]
run = 'fastapi dev mydatasales_back_end/app.py'
pre_test = 'task lint'
test = 'pytest --cov=mydatasales_back_end -vv'
post_test = 'coverage html'
lint = 'ruff check . ; ruff check . --diff'
format = 'ruff check . --fix ; ruff format'

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
ignr = "^2.2"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies] 
ruff = "^0.5.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[tool.ruff.lint]
preview = true
select = ['I','F','E','W','PL','PT']

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
